# docs/master_swagger.yaml
basePath: /api/
definitions:
...
  curd.ParamSearch:
    properties:
      filter:
        additionalProperties: true
        type: object
      keyword:
        additionalProperties:
          type: string
        type: object
      limit:
        type: integer
      skip:
        type: integer
      sort:
        additionalProperties:
          type: integer
        type: object
    type: object
...
  curd.ReplyData-int64:
    properties:
      data:
        type: integer
      error:
        type: string
    type: object
...
  curd.ReplyData-model_Subscription:
    properties:
      data:
        $ref: '#/definitions/types.Subscription'
      error:
        type: string
    type: object
  curd.ReplyData-model_User:
    properties:
      data:
        $ref: '#/definitions/types.User'
      error:
        type: string
    type: object
...
  curd.ReplyList-model_Subscription:
    properties:
      data:
        items:
          $ref: '#/definitions/types.Subscription'
        type: array
      error:
        type: string
      total:
        type: integer
    type: object
  curd.ReplyList-model_User:
    properties:
      data:
        items:
          $ref: '#/definitions/types.User'
        type: array
      error:
        type: string
      total:
        type: integer
    type: object
...
  types.Subscription:
    properties:
      channels:
        items:
          type: string
        type: array
      created:
        type: string
      device:
        type: string
      device_id:
        type: string
      disabled:
        description: 禁用
        type: boolean
      id:
        type: integer
      level:
        type: integer
      product:
        type: string
      product_id:
        type: string
      user_id:
        type: string
      validator_id:
        type: string
    type: object
  types.User:
    properties:
      created:
        type: string
      disabled:
        type: boolean
      email:
        type: string
      id:
        type: string
      name:
        type: string
      roles:
        items:
          type: string
        type: array
      username:
        type: string
    type: object
...
info:
  contact: {}
  description: API文档
  title: 物联大师接口文档
  version: 3.2 版本
paths:
...
  /subscription/{id}:
    get:
      consumes:
      - application/json
      description: 获取订阅
      parameters:
      - description: 订阅ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/curd.ReplyData-model_Subscription'
      summary: 获取订阅
      tags:
      - subscription
    post:
      consumes:
      - application/json
      description: 修改订阅
      parameters:
      - description: 订阅ID
        in: path
        name: id
        required: true
        type: integer
      - description: 订阅信息
        in: body
        name: subscription
        required: true
        schema:
          $ref: '#/definitions/types.Subscription'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/curd.ReplyData-model_Subscription'
      summary: 修改订阅
      tags:
      - subscription
  /subscription/{id}/delete:
    get:
      consumes:
      - application/json
      description: 删除订阅
      parameters:
      - description: 订阅ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/curd.ReplyData-model_Subscription'
      summary: 删除订阅
      tags:
      - subscription
  /subscription/{id}/disable:
    get:
      consumes:
      - application/json
      description: 禁用订阅
      parameters:
      - description: 订阅ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/curd.ReplyData-model_Subscription'
      summary: 禁用订阅
      tags:
      - subscription
  /subscription/{id}/enable:
    get:
      consumes:
      - application/json
      description: 启用订阅
      parameters:
      - description: 订阅ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/curd.ReplyData-model_Subscription'
      summary: 启用订阅
      tags:
      - subscription
  /subscription/count:
    post:
      consumes:
      - application/json
      description: 查询订阅数量
      parameters:
      - description: 查询参数
        in: body
        name: search
        required: true
        schema:
          $ref: '#/definitions/curd.ParamSearch'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/curd.ReplyData-int64'
      summary: 查询订阅数量
      tags:
      - subscription
  /subscription/create:
    post:
      consumes:
      - application/json
      description: 创建订阅
      parameters:
      - description: 订阅信息
        in: body
        name: search
        required: true
        schema:
          $ref: '#/definitions/types.Subscription'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/curd.ReplyData-model_Subscription'
      summary: 创建订阅
      tags:
      - subscription
...
  /subscription/import:
    post:
      consumes:
      - multipart/form-data
      description: 导入订阅
      parameters:
      - description: 压缩包
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/curd.ReplyData-int64'
      summary: 导入订阅
      tags:
      - subscription
  /subscription/list:
    get:
      consumes:
      - application/json
      description: 查询订阅
      parameters:
      - in: query
        name: limit
        type: integer
      - in: query
        name: skip
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/curd.ReplyList-model_Subscription'
      summary: 查询订阅
      tags:
      - subscription
  /subscription/search:
    post:
      consumes:
      - application/json
      description: 查询订阅
      parameters:
      - description: 查询参数
        in: body
        name: search
        required: true
        schema:
          $ref: '#/definitions/curd.ParamSearch'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/curd.ReplyList-model_Subscription'
      summary: 查询订阅
      tags:
      - subscription
  /user/{id}:
    get:
      consumes:
      - application/json
      description: 获取用户
      parameters:
      - description: 用户ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/curd.ReplyData-model_User'
      summary: 获取用户
      tags:
      - user
    post:
      consumes:
      - application/json
      description: 修改用户
      parameters:
      - description: 用户ID
        in: path
        name: id
        required: true
        type: string
      - description: 用户信息
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/types.User'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/curd.ReplyData-model_User'
      summary: 修改用户
      tags:
      - user
  /user/{id}/delete:
    get:
      consumes:
      - application/json
      description: 删除用户
      parameters:
      - description: 用户ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/curd.ReplyData-model_User'
      summary: 删除用户
      tags:
      - user
  /user/{id}/disable:
    get:
      consumes:
      - application/json
      description: 禁用用户
      parameters:
      - description: 用户ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/curd.ReplyData-model_User'
      summary: 禁用用户
      tags:
      - user
  /user/{id}/enable:
    get:
      consumes:
      - application/json
      description: 启用用户
      parameters:
      - description: 用户ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/curd.ReplyData-model_User'
      summary: 启用用户
      tags:
      - user
  /user/count:
    post:
      consumes:
      - application/json
      description: 查询用户数量
      parameters:
      - description: 查询参数
        in: body
        name: search
        required: true
        schema:
          $ref: '#/definitions/curd.ParamSearch'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/curd.ReplyData-int64'
      summary: 查询用户数量
      tags:
      - user
  /user/create:
    post:
      consumes:
      - application/json
      description: 创建用户
      parameters:
      - description: 用户信息
        in: body
        name: search
        required: true
        schema:
          $ref: '#/definitions/types.User'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/curd.ReplyData-model_User'
      summary: 创建用户
      tags:
      - user
  /user/list:
    get:
      consumes:
      - application/json
      description: 查询用户
      parameters:
      - in: query
        name: limit
        type: integer
      - in: query
        name: skip
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/curd.ReplyList-model_User'
      summary: 查询用户
      tags:
      - user
...
  /user/search:
    post:
      consumes:
      - application/json
      description: 查询用户
      parameters:
      - description: 查询参数
        in: body
        name: search
        required: true
        schema:
          $ref: '#/definitions/curd.ParamSearch'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/curd.ReplyList-model_User'
      summary: 查询用户
      tags:
      - user
...
swagger: "2.0"